⚙️ Refactor librarySlice
