+++ client/src/features/library/librarySlice.ts
@@ -1,4 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import processItems from './helpers/processItems.ts';
 
 interface Item {
   id: string;
@@ -24,42 +25,6 @@ const initialState: LibraryState = {
   sortBy: 'alphabetical',
   filter: 'none',
   searchQuery: ''
-};
-
-const processItems = (
-  items: Item[],
-  sortBy: 'alphabetical' | 'recentlyAdded',
-  filter: 'artist' | 'playlist' | 'none',
-  searchQuery: string
-): Item[] => {
-  let arr = [...items];
-
-  // 1. Sort items
-  arr = [...arr].sort((a, b) => {
-    const pinComparison = +b.isPinned - +a.isPinned;
-    if (pinComparison !== 0) return pinComparison;
-
-    if (sortBy === 'recentlyAdded') {
-      return +new Date(b.createdAt) - +new Date(a.createdAt);
-    } else {
-      // sort alphabetically by default
-      return a.name.localeCompare(b.name);
-    }
-  });
-
-  // 2. Filter items
-  arr =
-    filter !== 'none' ? arr.filter((item) => item.itemType === filter) : arr;
-
-  // 3. SearchPage items
-  arr =
-    searchQuery.length >= 3
-      ? arr.filter((item) =>
-        item.name.toLowerCase().includes(searchQuery.toLowerCase())
-      )
-      : arr;
-
-  return arr;
 };
 
 const librarySlice = createSlice({
